+++ src/main/java/com/spoony/spoony_server/domain/location/entity/LocationGuEntity.java
@@ -11,14 +11,14 @@ import lombok.NoArgsConstructor;
 public class LocationGuEntity {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer guId;
+    private Long guId;
     private String guName;
     private String guAddress;
     private Double latitude;
     private Double longitude;
 
     @Builder
-    public LocationGuEntity(Integer guId, String guName, String guAddress, Double latitude, Double longitude) {
+    public LocationGuEntity(Long guId, String guName, String guAddress, Double latitude, Double longitude) {
         this.guId = guId;
         this.guName = guName;
         this.guAddress = guAddress;
