+++ src/main/java/com/spoony/spoony_server/domain/report/repository/ReportRepository.java
@@ -3,7 +3,5 @@ package com.spoony.spoony_server.domain.report.repository;
 import com.spoony.spoony_server.domain.report.entity.ReportEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ReportRepository extends JpaRepository<ReportEntity, Integer> {
-
-
+public interface ReportRepository extends JpaRepository<ReportEntity, Long> {
 }
